.junie/scripts/requirements.txt
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
urllib3==2.5.0
